+++ package.json
@@ -80,7 +80,7 @@
     "gh-pages": "^1.2.0",
     "globby": "^8.0.1",
     "hard-rejection": "^1.0.0",
-    "highlight.js": "^9.12.0",
+    "highlight.js": "^11.6.0",
     "husky": "^4.3.8",
     "lint-staged": "^10.5.4",
     "lodash": "^4.17.21",
