--- src/components/code-snippet/examples/code-snippet-example-options.js
@@ -1,110 +1,0 @@
-/*
-Set maxHeight and onCopy
-*/
-import React from 'react';
-import CodeSnippet from '../code-snippet';
-
-export default class Example extends React.Component {
-  render() {
-    return (
-      <CodeSnippet
-        maxHeight={300}
-        onCopy={() => console.log('Copy!')}
-        code={`package com.mapbox.mapboxandroiddemo.examples.basics;
-
-import android.os.Bundle;
-
-import com.mapbox.mapboxandroiddemo.R;
-import com.mapbox.mapboxsdk.Mapbox;
-import com.mapbox.mapboxsdk.maps.MapView;
-import com.mapbox.mapboxsdk.maps.MapboxMap;
-import com.mapbox.mapboxsdk.maps.OnMapReadyCallback;
-import com.mapbox.mapboxsdk.maps.Style;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-
-
-/**
- * The most basic example of adding a map to an activity.
- */
-public class SimpleMapViewActivity extends AppCompatActivity {
-
-  private MapView mapView;
-
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-
-    // Mapbox access token is configured here. This needs to be called either in your application
-    // object or in the same activity which contains the mapview.
-    Mapbox.getInstance(this, getString(R.string.access_token));
-
-    // This contains the MapView in XML and needs to be called after the access token is configured.
-    setContentView(R.layout.activity_basic_simple_mapview);
-
-    mapView = findViewById(R.id.mapView);
-    mapView.onCreate(savedInstanceState);
-    mapView.getMapAsync(new OnMapReadyCallback() {
-      @Override
-      public void onMapReady(@NonNull MapboxMap mapboxMap) {
-        mapboxMap.setStyle(Style.MAPBOX_STREETS, new Style.OnStyleLoaded() {
-          @Override
-          public void onStyleLoaded(@NonNull Style style) {
-
-            // Map is set up and the style has loaded. Now you can add data or make other map adjustments.
-
-          }
-        });
-      }
-    });
-  }
-
-  // Add the mapView lifecycle to the activity's lifecycle methods
-  @Override
-  public void onResume() {
-    super.onResume();
-    mapView.onResume();
-  }
-
-  @Override
-  protected void onStart() {
-    super.onStart();
-    mapView.onStart();
-  }
-
-  @Override
-  protected void onStop() {
-    super.onStop();
-    mapView.onStop();
-  }
-
-  @Override
-  public void onPause() {
-    super.onPause();
-    mapView.onPause();
-  }
-
-  @Override
-  public void onLowMemory() {
-    super.onLowMemory();
-    mapView.onLowMemory();
-  }
-
-  @Override
-  protected void onDestroy() {
-    super.onDestroy();
-    mapView.onDestroy();
-  }
-
-  @Override
-  protected void onSaveInstanceState(Bundle outState) {
-    super.onSaveInstanceState(outState);
-    mapView.onSaveInstanceState(outState);
-  }
-}
-`}
-      />
-    );
-  }
-}
